requirements.txt
bit==0.8.0
python-dotenv==1.0.0
Requests==2.31.0
termcolor==2.3.0
